Create custom XHRResponse for ease of use
Use custom RmndinResponse as app.response_class that transforms the
XHRResponse into a proper response using the force_type classmethod.

This allows us to return a standardized response from the
controllers/endpoints.

The status that is passed in determines the status_code of the response
using the STATUS_MAP in rmndin.lib.web.response.

Example:
    return success(message="Success!", data=data, status="created")

    *success is just XHRSucces aliased in rmndin/__init__.py